+++ packages/to-json-schema/src/convertAction.ts
@@ -1,7 +1,7 @@
 import type { JSONSchema7 } from 'json-schema';
 import type * as v from 'valibot';
 import type { ConversionConfig } from './type.ts';
-import { handleError } from './utils/index.ts';
+import { addError, handleError } from './utils/index.ts';
 
 /**
  * Action type.
@@ -117,6 +117,10 @@ export function convertAction(
   valibotAction: Action,
   config: ConversionConfig | undefined
 ): JSONSchema7 {
+  // Create errors variable
+  let errors: [string, ...string[]] | undefined;
+
+  // Convert Valibot action to JSON Schema
   switch (valibotAction.type) {
     case 'base64': {
       jsonSchema.contentEncoding = 'base64';
@@ -152,9 +156,9 @@ export function convertAction(
         jsonSchema.maxItems = 0;
       } else {
         if (jsonSchema.type !== 'string') {
-          handleError(
-            `The "${valibotAction.type}" action is not supported on type "${jsonSchema.type}".`,
-            config
+          errors = addError(
+            errors,
+            `The "${valibotAction.type}" action is not supported on type "${jsonSchema.type}".`
           );
         }
         jsonSchema.maxLength = 0;
@@ -205,9 +209,9 @@ export function convertAction(
         jsonSchema.maxItems = valibotAction.requirement;
       } else {
         if (jsonSchema.type !== 'string') {
-          handleError(
-            `The "${valibotAction.type}" action is not supported on type "${jsonSchema.type}".`,
-            config
+          errors = addError(
+            errors,
+            `The "${valibotAction.type}" action is not supported on type "${jsonSchema.type}".`
           );
         }
         jsonSchema.minLength = valibotAction.requirement;
@@ -226,9 +230,9 @@ export function convertAction(
         jsonSchema.maxItems = valibotAction.requirement;
       } else {
         if (jsonSchema.type !== 'string') {
-          handleError(
-            `The "${valibotAction.type}" action is not supported on type "${jsonSchema.type}".`,
-            config
+          errors = addError(
+            errors,
+            `The "${valibotAction.type}" action is not supported on type "${jsonSchema.type}".`
           );
         }
         jsonSchema.maxLength = valibotAction.requirement;
@@ -238,9 +242,9 @@ export function convertAction(
 
     case 'max_value': {
       if (jsonSchema.type !== 'number') {
-        handleError(
-          `The "max_value" action is not supported on type "${jsonSchema.type}".`,
-          config
+        errors = addError(
+          errors,
+          `The "max_value" action is not supported on type "${jsonSchema.type}".`
         );
       }
       // @ts-expect-error
@@ -271,9 +275,9 @@ export function convertAction(
         jsonSchema.minItems = valibotAction.requirement;
       } else {
         if (jsonSchema.type !== 'string') {
-          handleError(
-            `The "${valibotAction.type}" action is not supported on type "${jsonSchema.type}".`,
-            config
+          errors = addError(
+            errors,
+            `The "${valibotAction.type}" action is not supported on type "${jsonSchema.type}".`
           );
         }
         jsonSchema.minLength = valibotAction.requirement;
@@ -283,9 +287,9 @@ export function convertAction(
 
     case 'min_value': {
       if (jsonSchema.type !== 'number') {
-        handleError(
-          `The "min_value" action is not supported on type "${jsonSchema.type}".`,
-          config
+        errors = addError(
+          errors,
+          `The "min_value" action is not supported on type "${jsonSchema.type}".`
         );
       }
       // @ts-expect-error
@@ -303,9 +307,9 @@ export function convertAction(
         jsonSchema.minItems = 1;
       } else {
         if (jsonSchema.type !== 'string') {
-          handleError(
-            `The "${valibotAction.type}" action is not supported on type "${jsonSchema.type}".`,
-            config
+          errors = addError(
+            errors,
+            `The "${valibotAction.type}" action is not supported on type "${jsonSchema.type}".`
           );
         }
         jsonSchema.minLength = 1;
@@ -315,7 +319,10 @@ export function convertAction(
 
     case 'regex': {
       if (valibotAction.requirement.flags) {
-        handleError('RegExp flags are not supported by JSON Schema.', config);
+        errors = addError(
+          errors,
+          'RegExp flags are not supported by JSON Schema.'
+        );
       }
       jsonSchema.pattern = valibotAction.requirement.source;
       break;
@@ -346,12 +353,33 @@ export function convertAction(
     }
 
     default: {
-      handleError(
+      errors = addError(
+        errors,
         // @ts-expect-error
-        `The "${valibotAction.type}" action cannot be converted to JSON Schema.`,
-        config
+        `The "${valibotAction.type}" action cannot be converted to JSON Schema.`
       );
     }
   }
+
+  // Override JSON Schema if specified and necessary
+  if (config?.overrideAction) {
+    const actionOverride = config.overrideAction({
+      valibotAction,
+      jsonSchema,
+      errors,
+    });
+    if (actionOverride) {
+      return { ...actionOverride };
+    }
+  }
+
+  // Handle errors based on configuration
+  if (errors) {
+    for (const message of errors) {
+      handleError(message, config);
+    }
+  }
+
+  // Return converted JSON Schema
   return jsonSchema;
 }