+++ packages/to-json-schema/src/index.ts
@@ -1,2 +1,6 @@
-export * from './toJsonSchema.ts';
+export type { JSONSchema7 } from 'json-schema';
 export type { ConversionConfig } from './type.ts';
+
+export * from './globalDefs.ts';
+export * from './toJsonSchema.ts';
+export * from './toJsonSchemaDefs.ts';